requirements.txt
# Core Django
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL support

# ML/AI
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
# tensorflow>=2.13.0  # Optional, uncomment if needed for advanced ML

# Blockchain (optional - for production blockchain integration)
# web3>=6.9.0  # Ethereum integration - uncomment if needed
# eth-account>=0.8.0  # Uncomment if needed

# AWS
boto3>=1.28.0
awscli>=1.29.0

# API & Authentication
djangorestframework-simplejwt>=5.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development
ipython>=8.14.0



